llm/requirements.txt
kafka-python==2.0.2
transformers==4.35.2
torch==2.1.1
tokenizers==0.15.0
pydantic==2.5.0
python-dotenv==1.0.0
aiokafka==0.8.1
openai==1.3.5
huggingface_hub==0.19.4